publish sdk docs

+++ bin/travis-prep-website.sh
@@ -17,6 +17,12 @@ if [ ! -d ./packages/apps-style/website/build/style ]; then
   exit 1
 fi
 
+if [ ! -d ./packages/apps-sdk/target/docs ]; then
+  echo "Missing: packages/apps-sdk/target/docs"
+  echo "Did you build first?"
+  exit 1
+fi
+
 if [ -d ./target/gh-pages ]; then
   rm -rf ./target/gh-pages
 fi
@@ -25,6 +31,7 @@ mkdir -p ./target/gh-pages
 
 mv ./packages/apps-components/target/storybook ./target/gh-pages/components
 mv ./packages/apps-style/website/build/style ./target/gh-pages/style
+mv ./packages/apps-sdk/target/docs ./target/gh-pages/sdk
 
 cat << EOF > ./target/gh-pages/index.html
 <html>
@@ -51,7 +58,11 @@ cat << EOF > ./target/gh-pages/index.html
                 </div>
 
                 <div class="pluginWrapper buttonWrapper">
-                  <a class="button" href="https://github.com/deskpro/apps-components">GitHub</a>
+                  <a class="button" href="./sdk/reference/">SDK</a>
+                </div>
+
+                <div class="pluginWrapper buttonWrapper">
+                  <a class="button" href="https://github.com/deskpro/apps-libs">GitHub</a>
                 </div>
               </div>
             </div>
@@ -65,5 +76,5 @@ cat << EOF > ./target/gh-pages/index.html
 EOF
 
 echo ""
-echo "gh-pages directory is ready for publishing"
+echo "target/gh-pages directory is ready for publishing"
 echo ""
